requirements.txt
python-dotenv==1.0.0
pydantic==2.5.2
pymongo[srv]==4.5.0
google-api-python-client==2.97.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
pytest
pandas

scikit-learn
scikit-fuzzy

torch
transformers[torch]
sentence-transformers[torch]

wandb

atpbar
tqdm

rich

fastapi==0.104.1

boto3

randomname
gensim
beir

pytest-mock